requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==5.1.5
django-crispy-forms==2.1
django-jazzmin==3.0.1
docopt==0.6.2
et_xmlfile==2.0.0
firebase==4.0.1
Flask==3.1.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
pymysql==1.1.0
mysqlclient==2.2.7
num2words==0.5.14
numpy==2.2.4
openai==1.75.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfkit==1.0.0
pefile==2023.2.7
pillow==10.4.0
pydantic==2.11.3
pydantic_core==2.33.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32-ctypes==0.2.3
reportlab==4.0.8
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.6.0
Flask==2.0.1
psycopg2-binary==2.9.3
Werkzeug==2.0.1
python-dotenv==0.19.0
gunicorn==20.1.0